requirements.txt
PySide6
psutil
GPUtil
py-cpuinfo
